+++ rectification/src/about.cpp
@@ -11,11 +11,10 @@
  *  The dialog will by default be modeless, unless you set 'modal' to
  *  true to construct a modal dialog.
  */
-About::About(QWidget* parent, const char* name, bool modal, Qt::WindowFlags fl)
-    : QDialog(parent, name, modal, fl)
+About::About(QWidget* parent)
+    : QDialog(parent)
 {
     setupUi(this);
-
 }
 
 /*
