+++ rectification/src/form2.cpp
@@ -1,10 +1,12 @@
 #include "form2.h"
 
-#include <qvariant.h>
+#include <q3filedialog.h>
+#include <Q3TextStream>
 #include <qimage.h>
+#include <qmessagebox.h>
 #include <qpixmap.h>
+#include <qvariant.h>
 
-#include "ui/form2.ui.h"
 /*
  *  Constructs a Form2 as a child of 'parent', with the
  *  name 'name' and widget flags set to 'f'.
@@ -16,7 +18,6 @@ Form2::Form2(QWidget* parent, const char* name, bool modal, Qt::WindowFlags fl)
     : QDialog(parent, name, modal, fl)
 {
     setupUi(this);
-
 }
 
 /*
@@ -36,3 +37,34 @@ void Form2::languageChange()
     retranslateUi(this);
 }
 
+void Form2::mensagem(QString s)
+{
+    textEdit1->insert(s);
+}
+
+void Form2::salvarImagem()
+{
+    QString filename = Q3FileDialog::getSaveFileName(".", "Text (*.txt)", this, "Save Report Dialog", "Choose one name to save Report text") ;
+    if (filename == "")
+        return;
+    // Verifica se arquivo ja existe
+    if ( QFile::exists( filename ) &&
+            QMessageBox::warning(
+                this,
+                tr("Warning: Overwrite File?"),
+                tr("A file called %1 already exists. "
+                   "Do you want to overwrite it?")
+                .arg( filename ),
+                tr("&Yes"), tr("&No"),
+                QString::null, 0, 1 ) )
+        return;
+
+    QString text = textEdit1->text();
+    QFile f( filename );
+    if ( !f.open( QIODevice::WriteOnly ) )
+        return;
+
+    Q3TextStream t( &f );
+    t << text;
+    f.close();
+}
