+++ rectification/ui/form1.ui
@@ -347,7 +347,7 @@
         </property>
        </item>
       </widget>
-      <widget class="QComboBox" name="comboBox1">
+      <widget class="QComboBox" name="transformation">
        <property name="geometry">
         <rect>
          <x>10</x>
@@ -355,6 +355,9 @@
          <width>180</width>
          <height>21</height>
         </rect>
+       </property>
+       <property name="currentIndex">
+        <number>5</number>
        </property>
        <item>
         <property name="text">
@@ -548,7 +551,7 @@
           </font>
          </property>
          <property name="text">
-          <string>Destiny:</string>
+          <string>Destination:</string>
          </property>
          <property name="alignment">
           <set>Qt::AlignCenter</set>
@@ -597,6 +600,9 @@
              </property>
              <property name="maximum">
               <number>20</number>
+             </property>
+             <property name="value">
+              <number>4</number>
              </property>
             </widget>
            </item>
@@ -1330,7 +1336,7 @@
   <tabstop>YiEdit</tabstop>
   <tabstop>XfEdit</tabstop>
   <tabstop>YfEdit</tabstop>
-  <tabstop>comboBox1</tabstop>
+  <tabstop>transformation</tabstop>
   <tabstop>interpolation</tabstop>
   <tabstop>XoEdit</tabstop>
   <tabstop>YoEdit</tabstop>
@@ -1368,7 +1374,7 @@
    </hints>
   </connection>
   <connection>
-   <sender>comboBox1</sender>
+   <sender>transformation</sender>
    <signal>activated(QString)</signal>
    <receiver>Form1</receiver>
    <slot>minimosPontos()</slot>
